requirements.txt
requests
crossplane
pyyaml